LayoutTests/fast/forms/ValidityState-stepMismatch-expected.txt

Check stepMismatch results for type=date, datetime, datetime-local, month, time, week.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Date type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "2", "1969-12-31") is false
Normal step values
PASS stepMismatchFor("2010-02-10", "2", "2010-02-09") is true
PASS stepMismatchFor("2010-02-09", "2", "2010-02-09") is false
PASS stepMismatchFor("2010-02-11", "2", "2010-02-09") is false
PASS stepMismatchFor("1800-11-11", "3", "1800-11-09") is true
PASS stepMismatchFor("1800-11-09", "3", "1800-11-09") is false
PASS stepMismatchFor("1800-11-12", "3", "1800-11-09") is false
PASS stepMismatchFor("275760-09-13", "3", "275760-09-11") is true
PASS stepMismatchFor("275760-09-13", "2", "275760-09-11") is false
Implicit step base
PASS stepMismatchFor("1970-01-02", "2", null) is true
PASS stepMismatchFor("1970-01-03", "2", null) is false
Fractional step values
PASS stepMismatchFor("2010-02-10", "0.1", "2010-02-09") is false
PASS stepMismatchFor("2010-02-10", "1.1", "2010-02-09") is false
PASS stepMismatchFor("2010-02-10", "1.9", "2010-02-09") is true
Invalid or no step values
PASS stepMismatchFor("2010-02-10", null, "2010-02-09") is false
PASS stepMismatchFor("2010-02-10", "-1", "2010-02-09") is false
PASS stepMismatchFor("2010-02-10", "foo", "2010-02-09") is false
Special step value
PASS stepMismatchFor("2010-02-10", "any", "2010-02-09") is false

Datetime type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "2", "1969-12-31T12:34:56Z") is false
Normal step values
PASS stepMismatchFor("2010-02-09T12:34:55Z", "1", "2010-02-09T12:34:56Z") is false
PASS stepMismatchFor("2010-02-09T12:34:55.001Z", "1", "2010-02-09T12:34:56Z") is true
PASS stepMismatchFor("2010-02-09T12:34:56.001Z", "0.001", "2010-02-09T12:34:56Z") is false
PASS stepMismatchFor("2010-02-09T12:34:55Z", "0.333", "2010-02-09T12:34:56Z") is true
PASS stepMismatchFor("2010-02-09T12:34:55.001Z", "0.333", "2010-02-09T12:34:56Z") is false
PASS stepMismatchFor("2010-02-09T12:34Z", "86400", "2010-02-08T12:34Z") is false
PASS stepMismatchFor("2010-02-09T12:34:56Z", "86400", "2010-02-08T12:34Z") is true
PASS stepMismatchFor("275760-09-13T00:00Z", "3", "275760-09-12T23:59:50Z") is true
PASS stepMismatchFor("275760-09-13T00:00Z", "2", "275760-09-12T23:59:50Z") is false
PASS stepMismatchFor("0001-01-01T00:00Z", "3", "0001-01-11T00:00:02Z") is true
PASS stepMismatchFor("0001-01-01T00:00Z", "2", "0001-01-11T00:00:02Z") is false
Implicit step base
PASS stepMismatchFor("1970-01-01T12:34Z", "120", null) is false
PASS stepMismatchFor("1970-01-01T12:35Z", "120", null) is true
Small step values
PASS stepMismatchFor("2010-02-10T12:34:56.000Z", "0.0003", "2010-02-10T12:34.55.000Z") is false
PASS stepMismatchFor("2010-02-10T12:34:55.001Z", "0.0019", "2010-02-10T12:34.55.000Z") is true
Invalid or no step values
PASS stepMismatchFor("2010-02-10T12:34Z", null, "2010-02-09T12:34Z") is false
PASS stepMismatchFor("2010-02-10T12:34:56Z", null, "2010-02-09T12:34Z") is true
PASS stepMismatchFor("2010-02-10T12:34Z", "-1", "2010-02-09T12:34Z") is false
PASS stepMismatchFor("2010-02-10T12:34Z", "foo", "2010-02-09T12:34Z") is false
Special step value
PASS stepMismatchFor("2010-02-09T12:34Z", "any", "2010-02-09T12:34Z") is false

Datetime-local type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "2", "1969-12-31T12:34:56") is false
Normal step values
PASS stepMismatchFor("2010-02-09T12:34:55", "1", "2010-02-09T12:34:56") is false
PASS stepMismatchFor("2010-02-09T12:34:55.001", "1", "2010-02-09T12:34:56") is true
PASS stepMismatchFor("2010-02-09T12:34:56.001", "0.001", "2010-02-09T12:34:56") is false
PASS stepMismatchFor("2010-02-09T12:34:55", "0.333", "2010-02-09T12:34:56") is true
PASS stepMismatchFor("2010-02-09T12:34:55.001", "0.333", "2010-02-09T12:34:56") is false
PASS stepMismatchFor("2010-02-09T12:34", "86400", "2010-02-08T12:34") is false
PASS stepMismatchFor("2010-02-09T12:34:56", "86400", "2010-02-08T12:34") is true
PASS stepMismatchFor("275760-09-13T00:00", "3", "275760-09-12T23:59:50") is true
PASS stepMismatchFor("275760-09-13T00:00", "2", "275760-09-12T23:59:50") is false
PASS stepMismatchFor("0001-01-01T00:00", "3", "0001-01-11T00:00:02") is true
PASS stepMismatchFor("0001-01-01T00:00", "2", "0001-01-11T00:00:02") is false
Implicit step base
PASS stepMismatchFor("1970-01-01T12:34", "120", null) is false
PASS stepMismatchFor("1970-01-01T12:35", "120", null) is true
Small step values
PASS stepMismatchFor("2010-02-10T12:34:56.000", "0.0003", "2010-02-10T12:34.55.000") is false
PASS stepMismatchFor("2010-02-10T12:34:55.001", "0.0019", "2010-02-10T12:34.55.000") is true
Invalid or no step values
PASS stepMismatchFor("2010-02-10T12:34", null, "2010-02-09T12:34") is false
PASS stepMismatchFor("2010-02-10T12:34:56", null, "2010-02-09T12:34") is true
PASS stepMismatchFor("2010-02-10T12:34", "-1", "2010-02-09T12:34") is false
PASS stepMismatchFor("2010-02-10T12:34", "foo", "2010-02-09T12:34") is false
Special step value
PASS stepMismatchFor("2010-02-09T12:34", "any", "2010-02-09T12:34") is false

Month type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "2", "1969-12") is false
Normal step values
PASS stepMismatchFor("2010-03", "2", "2010-02") is true
PASS stepMismatchFor("2010-02", "2", "2010-02") is false
PASS stepMismatchFor("2010-04", "2", "2010-02") is false
PASS stepMismatchFor("1800-11", "3", "1800-09") is true
PASS stepMismatchFor("1800-09", "3", "1800-09") is false
PASS stepMismatchFor("1800-12", "3", "1800-09") is false
PASS stepMismatchFor("275760-09", "3", "275760-08") is true
PASS stepMismatchFor("275760-09", "2", "275760-09") is false
Implicit step base
PASS stepMismatchFor("1970-02", "2", null) is true
PASS stepMismatchFor("1970-03", "2", null) is false
Fractional step values
PASS stepMismatchFor("2010-03", "0.1", "2010-02") is false
PASS stepMismatchFor("2010-03", "1.1", "2010-02") is false
PASS stepMismatchFor("2010-03", "1.9", "2010-02") is true
Invalid or no step values
PASS stepMismatchFor("2010-03", null, "2010-02") is false
PASS stepMismatchFor("2010-03", "-1", "2010-02") is false
PASS stepMismatchFor("2010-03", "foo", "2010-02") is false
Special step value
PASS stepMismatchFor("2010-03", "any", "2010-02") is false

Number type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "1.0", "0.1") is false
Integers
PASS stepMismatchFor("1", "2", "0") is true
PASS stepMismatchFor("-3", "2", "-4") is true
PASS input.max = "5"; stepMismatchFor("5", "3", "0") is true
PASS input.value is "5"
Invalid step values
PASS stepMismatchFor("-3", "-2", "-4") is false
PASS stepMismatchFor("-3", null, "-4") is false
PASS stepMismatchFor("-3", undefined, "-4") is false
Huge numbers and small step; uncomparable
PASS stepMismatchFor("1.7976931348623157e+308", "3", "") is false
PASS stepMismatchFor("1.7976931348623156e+308", "3", "") is false
PASS stepMismatchFor("1.7976931348623155e+308", "3", "") is false
Huge numbers and huge step
PASS stepMismatchFor("1.60e+308", "0.20e+308", "") is false
PASS stepMismatchFor("1.60e+308", "0.22e+308", "") is true
Fractional numbers
PASS stepMismatchFor("0.9", "0.1", "") is false
PASS stepMismatchFor("0.9", "0.1000001", "") is true
PASS stepMismatchFor("0.9", "0.1000000000000001", "") is false
PASS stepMismatchFor("1.0", "0.3333333333333333", "") is false

Range type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "1.0", "0.1") is false
Integers
PASS stepMismatchFor("1", "2", "0") is false
PASS stepMismatchFor("-3", "2", "-4") is false
PASS input.max = "5"; stepMismatchFor("5", "3", "0") is false
PASS input.value is "3"
Invalid step values
PASS stepMismatchFor("-3", "-2", "-4") is false
PASS stepMismatchFor("-3", null, "-4") is false
PASS stepMismatchFor("-3", undefined, "-4") is false
Huge numbers and small step; uncomparable
PASS stepMismatchFor("1.7976931348623157e+308", "3", "") is false
PASS stepMismatchFor("1.7976931348623156e+308", "3", "") is false
PASS stepMismatchFor("1.7976931348623155e+308", "3", "") is false
Huge numbers and huge step
PASS stepMismatchFor("1.60e+308", "0.20e+308", "") is false
PASS stepMismatchFor("1.60e+308", "0.22e+308", "") is false
Fractional numbers
PASS stepMismatchFor("0.9", "0.1", "") is false
PASS stepMismatchFor("0.9", "0.1000001", "") is false
PASS stepMismatchFor("0.9", "0.1000000000000001", "") is false
PASS stepMismatchFor("1.0", "0.3333333333333333", "") is false

Time type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "2", "12:34:56") is false
Normal step values
PASS stepMismatchFor("12:34:55", "1", "12:34:56") is false
PASS stepMismatchFor("12:34:55.001", "1", "12:34:56") is true
PASS stepMismatchFor("12:34:56.001", "0.001", "12:34:56") is false
PASS stepMismatchFor("12:34:55", "0.333", "12:34:56") is true
PASS stepMismatchFor("12:34:55.001", "0.333", "12:34:56") is false
PASS stepMismatchFor("12:34", "86400", "12:34") is false
PASS stepMismatchFor("23:59:59.998", "86399.999", "00:00") is true
PASS stepMismatchFor("23:59:59.999", "86399.999", "00:00") is false
Implicit step base
PASS stepMismatchFor("12:34", "120", null) is false
PASS stepMismatchFor("12:35", "120", null) is true
Small step values
PASS stepMismatchFor("12:34:55.001", "0.0003", "12:34.55.000") is false
PASS stepMismatchFor("12:34:55.001", "0.0019", "12:34.55.000") is true
PASS stepMismatchFor("12:34:55.004", "0.0019", "12:34.55.000") is false
Invalid or no step values
PASS stepMismatchFor("12:34", null, "12:34") is false
PASS stepMismatchFor("12:34:56", null, "12:34") is true
PASS stepMismatchFor("12:34", "-1", "12:34") is false
PASS stepMismatchFor("12:34", "foo", "12:34") is false
Special step value
PASS stepMismatchFor("12:35", "any", "12:34") is false

Week type
Empty values
PASS stepMismatchFor("", null, null) is false
PASS stepMismatchFor("", "2", "1970-W02") is false
Normal step values
PASS stepMismatchFor("2010-W03", "2", "2010-W02") is true
PASS stepMismatchFor("2010-W02", "2", "2010-W02") is false
PASS stepMismatchFor("2010-W04", "2", "2010-W02") is false
PASS stepMismatchFor("1800-W11", "3", "1800-W09") is true
PASS stepMismatchFor("1800-W09", "3", "1800-W09") is false
PASS stepMismatchFor("1800-W12", "3", "1800-W09") is false
PASS stepMismatchFor("275760-W35", "3", "275760-W33") is true
PASS stepMismatchFor("275760-W35", "2", "275760-W33") is false
Implicit step base
PASS stepMismatchFor("1970-W01", "2", null) is false
PASS stepMismatchFor("1970-W02", "2", null) is true
PASS stepMismatchFor("1970-W03", "2", null) is false
PASS stepMismatchFor("1970-W04", "2", null) is true
Fractional step values
PASS stepMismatchFor("2010-W03", "0.1", "2010-W02") is false
PASS stepMismatchFor("2010-W03", "1.1", "2010-W02") is false
PASS stepMismatchFor("2010-W03", "1.9", "2010-W02") is true
Invalid or no step values
PASS stepMismatchFor("2010-W03", null, "2010-W02") is false
PASS stepMismatchFor("2010-W03", "-1", "2010-W02") is false
PASS stepMismatchFor("2010-W03", "foo", "2010-W02") is false
Special step value
PASS stepMismatchFor("2010-W03", "any", "2010-W02") is false

Unsupported types
PASS input.type = "text"; input.step = "3"; input.min = ""; input.value = "2"; input.validity.stepMismatch is false
PASS input.type = "button"; input.validity.stepMismatch is false
PASS input.type = "checkbox"; input.validity.stepMismatch is false
PASS input.type = "color"; input.validity.stepMismatch is false
PASS input.type = "email"; input.validity.stepMismatch is false
PASS input.type = "hidden"; input.validity.stepMismatch is false
PASS input.type = "image"; input.validity.stepMismatch is false
PASS input.type = "khtml_isindex"; input.validity.stepMismatch is false
PASS input.type = "passwd"; input.validity.stepMismatch is false
PASS input.type = "radio"; input.validity.stepMismatch is false
PASS input.type = "reset"; input.validity.stepMismatch is false
PASS input.type = "search"; input.validity.stepMismatch is false
PASS input.type = "submit"; input.validity.stepMismatch is false
PASS input.type = "tel"; input.validity.stepMismatch is false
PASS input.type = "url"; input.validity.stepMismatch is false
PASS input.type = "file"; input.validity.stepMismatch is false
PASS successfullyParsed is true

TEST COMPLETE
